Remove useless headers
